Move Views associated to emojis into a new plugin
