+++ src/templates/emojis.html
@@ -10,9 +10,9 @@
     </div>
     <div class="emoji-picker__lists">
         {[ Object.keys(o.emoji_categories).forEach(function (category) { ]}
-        <a id="emoji-picker-{{{category}}}" class="emoji-category__heading">{{{o._converse.emoji_category_labels[category]}}}</a>
+            <a id="emoji-picker-{{{category}}}" class="emoji-category__heading">{{{o._converse.emoji_category_labels[category]}}}</a>
             <ul class="emoji-picker emoji-picker-{{{category}}}">
-                {[ o._.forEach(o.emojis_by_category[category], function (emoji) { ]}
+                {[ o.emojis_by_category[category].forEach(function (emoji) { ]}
                 <li class="emoji insert-emoji {[ if (o.shouldBeHidden(emoji.sn, o.current_skintone, o.toned_emojis)) { ]} hidden {[ }; ]}"
                     data-emoji="{{{emoji.sn}}}" title="{{{emoji.sn}}}">
                         <a href="#" data-emoji="{{{emoji.sn}}}"> {{ o.transform(emoji.sn) }}  </a>
@@ -21,11 +21,14 @@
             </ul>
         {[ }); ]}
     </div>
-    <ul class="emoji-skintone-picker">
-        {[ o._.forEach(o.skintones, function (skintone) { ]}
-            <li data-skintone="{{{skintone}}}" class="emoji-skintone {[ if (o.current_skintone === skintone) { ]} picked {[ } ]}">
-                <a class="pick-skintone" href="#" data-skintone="{{{skintone}}}"> {{ o.transform(':'+skintone+':') }} </a>
-            </li>
-        {[ }); ]}
-    </ul>
+    <div class="emoji-skintone-picker">
+        <label>Skin tone:</label>
+        <ul>
+            {[ o.skintones.forEach(function (skintone) { ]}
+                <li data-skintone="{{{skintone}}}" class="emoji-skintone {[ if (o.current_skintone === skintone) { ]} picked {[ } ]}">
+                    <a class="pick-skintone" href="#" data-skintone="{{{skintone}}}"> {{ o.transform(':'+skintone+':') }} </a>
+                </li>
+            {[ }); ]}
+        </ul>
+    </div>
 </div>
